frontend/web/requirements.txt
gradio>=5.30.0
python-dotenv>=1.0.0
requests>=2.31.0
PyPDF2==3.0.1
